FD37 | feat(routes): export controllers
