+++ server/routes/api/restaurants/categories/index.js
@@ -1,15 +1,6 @@
 const router = require('express').Router();
-const {
-  addRestaurantCategory,
-} = require('../../../../db/actions/restaurant/category');
+const restaurantCategoryController = require('../../../../db/controllers/restaurantCategories');
 
-const createCategory = (req, res) => {
-  // add category
-  addRestaurantCategory(req.body)
-    .then((doc) => res.status(200).json({ category: doc }))
-    .catch((err) => res.status(400).json({ error: err.message }));
-};
-
-router.post('/create', createCategory);
+router.post('/create', restaurantCategoryController.create);
 
 module.exports = router;
